Améliorer la gestion des parenthèses et des nombres négatifs dans le calcul; refactoriser la méthode tokenize et ajuster le formatage des résultats

+++ src/main/java/fr/tse/fise2/calculator/CalculationResult.java
@@ -25,10 +25,10 @@ public class CalculationResult {
      * @return Le résultat formaté.
      */
     public String getFormattedResult() {
-        if (result == (int) result) {
-            return String.valueOf((int) result);
-        } else if (result == - ((int) result)) {
-            return String.valueOf(- ((int) result));
+        if (result == (long) result) {
+            return String.valueOf((long) result);
+        } else if (result == - ((long) result)) {
+            return String.valueOf(- ((long) result));
         }
         return String.valueOf(result);
     }
